Feat: reverse, reverseK; Fix: ascend, descend
Swap the node's prev and next while traverse the queue to complete
reverse. reverseK seems not allow using malloc, so I create a
function by myself to reverse k nodes while traversing the queue.

Fixing ascend and descend with variable name and if condition.
Variable name changing is because I forgot to change it while
copy paste. Condition is because the value have to be "strictly"
less or greater.