Feat: q_merge, Fix: q_descend
In q_merge, we traverse the queues whose head is chained in chain,
and splice from the second queue to the last queue to the tail of
first queue, and use q_sort to sort the merged queue.

q_descend occured segmentation fault and memory leak problem.
Segmentation fault is because when using list_for_each_entry_safe,
it will also do strcmp while &safe->list is equal to head.
Memory leak is because the description of q_descend is 'remove' the
element from the list, but it should be 'delete'.After adding
q_release_element then the problem is solved.